requirements.txt
flask==2.3.3
google-generativeai==0.3.1
tiktoken==0.5.1
gunicorn==21.2.0
flask-cors==4.0.0
python-dotenv==1.0.0 